Change and cleanup deprecated functions
Added a class Deprecated and moved old functions there, these are being kept forreference. Cleaned up and fixed the buildSentence <Single> and <Plural> Multiplefunctions to work correctly.
The old getRandomFileLine is deprecated and references to this function were renamed to getRandomFileLine_dep
Added and fixed words-sat-nouns to contain only nouns